Keep the dev React debug channel off the render layer's React imports
The dev-only React debug channel regressed with Node streams enabled:
`connectReactDebugChannel` pulled `createNodeBufferedTransformStream`
from `stream-ops.node`, which carries unconditional `react-dom/server`
and `react-dom/static` value imports. The hot-reloaders run in the
precompiled dev-server runtime, where `react-dom/*` is not aliased to
the vendored React, so that `require` resolved `react-dom/static`
against the app's installed React. On React 18 there is no `./static`
subpath export, which threw `ERR_PACKAGE_PATH_NOT_EXPORTED` and broke
app rendering in dev.

This change extracts the Node byte batcher into its own React-free
module, `stream-utils/node-buffered-transform-stream`, that depends only
on `node:stream`. Both `stream-ops.node` and the debug channel import it
from there, so the dev server can batch Node streams without dragging
the render layer's React imports into an unaliased runtime. The debug
channel's runtime branch on the stream type is unchanged, and the web
path already used a React-free batcher, so no `.node`/`.web` split is
needed.
